+++ CHANGELOG.md
@@ -35,9 +35,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
  - [#189](https://github.com/babylonlabs-io/btc-staker/pull/189) Use signing context
  string in staker pop
  - [#188](https://github.com/babylonlabs-io/btc-staker/pull/188) Add stake expansion feature
+ - [#190](https://github.com/babylonlabs-io/btc-staker/pull/190) Add UTXO consolidation command
 
 ## v0.16.0-rc.0